Move vocabulary related schemas into submodule.
